+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Planned changes/updates/fixes/features for to be part of one of the next _releases/versions_:
   * ...
 
+## [0.3.0] - 2021-0m-xx
+### Added
+  * Option to write request response to file
+
+### Fixed
+  * Fixed wrong `config-template.yaml`  
+  * Fixed adding headers if no headers set as parameters
+
 ## [0.2.0] - 2021-04-02
 ### Added
   * Support for headers in configuration files